add "jest": {     "coverageDirectory": "./coverage/",     "collectCoverage": true   },
