+++ package.json
@@ -3,6 +3,10 @@
   "version": "0.0.1",
   "description": "ITGM app",
   "main": "index.js",
+  "jest": {
+    "coverageDirectory": "./coverage/",
+    "collectCoverage": true
+  },
   "scripts": {
     "start": "parcel index.html",
     "build": "NODE_ENV=production parcel build index.html",
